Trying out some vibe-based perf.

The first commit here made this effect locally (removed compression benchmarks):
```
dbtext/wikipedia/decompress
                        time:   [887.44 µs 899.12 µs 911.18 µs]
                        thrpt:  [2.9261 GiB/s 2.9654 GiB/s 3.0044 GiB/s]
                 change:
                        time:   [-11.234% -10.036% -8.7806%] (p = 0.00 < 0.05)
                        thrpt:  [+9.6258% +11.156% +12.656%]
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild

dbtext/l_comment/decompress
                        time:   [410.96 µs 414.94 µs 418.83 µs]
                        thrpt:  [6.1060 GiB/s 6.1632 GiB/s 6.2228 GiB/s]
                 change:
                        time:   [-1.0256% +0.1567% +1.4423%] (p = 0.80 > 0.05)
                        thrpt:  [-1.4218% -0.1564% +1.0362%]
                        No change in performance detected.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

dbtext/urls/decompress  time:   [1.1473 ms 1.1569 ms 1.1675 ms]
                        thrpt:  [5.0477 GiB/s 5.0940 GiB/s 5.1368 GiB/s]
                 change:
                        time:   [+0.5443% +1.9998% +3.5034%] (p = 0.01 < 0.05)
                        thrpt:  [-3.3848% -1.9606% -0.5414%]
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild
```

Second one is more x86 oriented, so I'm going to let codspeed take a look at it.